automation-framework/requirements.txt
# ─────────────────────────────────────────────
# Core Framework
# ─────────────────────────────────────────────
robotframework>=7.4.1
robotframework-requests>=0.9.7

# ─────────────────────────────────────────────
# Web Automation
# ─────────────────────────────────────────────
robotframework-browser>=19.0.0
playwright>=1.58.0
playwright-stealth>=2.0.0
robotframework-seleniumlibrary>=6.8.0
selenium>=4.0.0

# ─────────────────────────────────────────────
# Mobile Automation
# ─────────────────────────────────────────────
robotframework-appiumlibrary>=3.2.1
Appium-Python-Client>=5.2.0

# ─────────────────────────────────────────────
# API & HTTP
# ─────────────────────────────────────────────
requests>=2.32.0
httpx>=0.28.0

# ─────────────────────────────────────────────
# Configuration & Environment
# ─────────────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.2.0

# ─────────────────────────────────────────────
# Validation
# ─────────────────────────────────────────────
jsonschema>=4.0

# ─────────────────────────────────────────────
# Load Testing
# ─────────────────────────────────────────────
locust>=2.43.0

# ─────────────────────────────────────────────
# Development & Testing
# ─────────────────────────────────────────────
pytest>=9.0.0

# ─────────────────────────────────────────────
# Database (optional, for data verification)
# ─────────────────────────────────────────────
robotframework-databaselibrary>=2.4.0
# psycopg2-binary>=2.9
# pymongo>=4.0
# mysql-connector-python>=8.0
